requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=3.14
PyJWT>=2.8
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
psycopg2-binary>=2.9
django-mongodb-backend>=0.1.0
# MongoDB Driver
pymongo>=4.6.0
# DNS resolver para MongoDB Atlas (conexões mongodb+srv://)
dnspython>=2.4.0
# Carregar variáveis de ambiente de arquivos .env
python-dotenv>=1.0.0
# Google OAuth2 - verificação de ID Token
google-auth>=2.28.0
# Cache de certificados do Google para verify_oauth2_token
cachecontrol>=0.14.0

# Google API client libraries for Gmail OAuth/send
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.70.0



